pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anvaya-voice-agent"
version = "0.1.0"
description = "Anvaya AI voice agent with Whisper transcription and Groq chat integration"
readme = "README.md"
authors = [{name = "Anvaya"}]
license = {text = "Proprietary"}
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0",
    "gradio>=4.44,<5",
    "faster-whisper>=1.0,<2",
    "groq>=0.11,<1",
    "pyttsx3>=2.90,<3",
    "soundfile>=0.12,<1",
    "numpy>=1.24",
    "typer[all]>=0.12,<1",
    "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
anvaya-agent = "anvaya_voice_agent.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.9"
packages = ["anvaya_voice_agent"]
ignore_missing_imports = true
